refactor: separate bitmap class into dedicated file
Move bitmap class definition from init.lua to bitmap.lua for better
code organization and separation of concerns.

Changes:
- Create src/pkg/bitmap/bitmap.lua with bitmap class and helper functions
- Update init.lua to just require and export the bitmap class
- No API changes - tests continue to work without modification

This makes the codebase more modular and easier to maintain.
